Work.

+++ df3d/engine/render/MeshUtils.h
@@ -10,11 +10,11 @@ class MeshUtils
 {
 public:
     static void indexize(const Vertex_p_n_tx_tan_bitan *vdata, size_t count,
-                         PodArray<Vertex_p_n_tx_tan_bitan> &outVertices, PodArray<uint32_t> &outIndices);
+                         PodArray<Vertex_p_n_tx_tan_bitan> &outVertices, PodArray<uint16_t> &outIndices);
 
     static void computeTangentBasis(Vertex_p_n_tx_tan_bitan *vdata, size_t count);
     static void computeTangentBasis(Vertex_p_n_tx_tan_bitan *vdata, size_t verticesCount,
-                                    const uint32_t *indices, size_t indicesCount);
+                                    const uint16_t *indices, size_t indicesCount);
 };
 
 }
